Rename : Move Auth Data Model Files' Directory
